refactor(HttpHeaders): HttpHeader에 대한 컨테이너 객체 구현
